% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_release.R
\name{plot_release}
\alias{plot_release}
\title{Plot Release Over Time}
\usage{
plot_release(release)
}
\arguments{
\item{release}{Numeric vector of release amounts over time.}
}
\value{
A ggplot object displaying the release over time.
}
\description{
This function plots the release data over time.
}
\examples{
plot_release(result$release)
}
