% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_data.R
\docType{data}
\name{reservoir_data}
\alias{reservoir_data}
\title{Reservoir Data}
\format{
A data frame with 12 rows and 5 variables:
\describe{
\item{Month}{Abbreviated month names (character).}
\item{Inflow}{Monthly inflow volumes in cubic meters (numeric).}
\item{Demand}{Monthly demand volumes in cubic meters (numeric).}
\item{S_min}{Minimum storage limits in cubic meters (numeric).}
\item{S_max}{Maximum storage limits in cubic meters (numeric).}
}
}
\source{
Mahaweli Authority of Sri Lanka.
}
\usage{
data(reservoir_data)
}
\description{
Monthly reservoir inflows, demands, and storage limits.
}
\details{
This dataset is used to demonstrate the optimization of reservoir releases using the \code{OptGoal} package.
}
\keyword{datasets}
