% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_release_vs_demand.R
\name{plot_release_vs_demand}
\alias{plot_release_vs_demand}
\title{Plot Release and Demand Over Time}
\usage{
plot_release_vs_demand(release, demand, months = NULL)
}
\arguments{
\item{release}{Numeric vector of release amounts over time.}

\item{demand}{Numeric vector of demand amounts over time.}

\item{months}{Optional character vector of month names corresponding to the data.
If not provided and the length of data is 12, defaults to Jan-Dec.}
}
\value{
A ggplot object displaying the release and demand over time.
}
\description{
This function plots both the release and demand data over time on the same graph,
allowing for easy comparison between the two.
}
\examples{
# Example usage
inflow <- c(7569, 6383, 8472, 13084, 9689,
            4301, 399, 2896, 2723, 7239, 14849, 13304)
demand <- c(1586.72, 1057.81, 1586.72, 1057.81, 1586.72,
            1057.81, 1586.72, 1057.81, 1586.72, 1057.81, 1586.72, 1057.81)
S_min_vector <- rep(2824.55, times = 12)
S_max_vector <- rep(26864.25, times = 12)
constraints <- list(
  initial_storage = 500,
  S_min = S_min_vector,
  S_max = S_max_vector
)
result <- optimize_release(inflow, demand, constraints)
release <- result$release
plot_release_vs_demand(release, demand)
}
